change the util method to RSA class
change the util method to RSA class
